Inicio de sesion con Sweet Alert

+++ js/inicio.js
@@ -17,49 +17,46 @@ let pw;
 */
 let verificar;
 
-const USER = document.getElementById("inicio-usuario");
-const PW = document.getElementById("inicio-pw")
-document.getElementById("validar").addEventListener('click', function IniciarSesion(event) {
-    event.preventDefault();
-    let user = USER.value;
-    let pw = PW.value;
-    verificar = !!USUARIOS.find(persona => persona.usuario === user && persona.contrasenia === pw);
-    if (verificar) {
-        console.log("Bienvenido/a" + " " + user);
-        localStorage.setItem("logCorrecto", true);
-        window.location.href = "/index.html";
-    }
-    else {
-        const ERROR = document.getElementById("error");
-        ERROR.textContent = "Usuario y/o contraseña erroneos, por favor, reintente"
 
-    }
 
-}
-)
+
 window.addEventListener("load", function () {
     if (localStorage.getItem("logCorrecto")) {
         window.location.href = "/index.html";
     }
 })
 
-const MODAL = document.getElementById("inicio-dialog");
-//Añadir llamado
-const PARAINGRESO = document.getElementsByClassName("iniciar-sesion");
-Array.from(PARAINGRESO).forEach(boton => {
-    boton.addEventListener("click", () => {
-        MODAL.showModal();
-    });
-})
 
-const CERRAR = document.getElementById("inicio-cancelar");
-CERRAR.addEventListener("click", () => {
-    MODAL.close();
-});
-//Añadir llamado
-const RINGRESOS = document.getElementsByClassName("iniciar-sesion");
-Array.from(RINGRESOS).forEach(boton => {
-    boton.addEventListener("click", () => {
-        MODAL.showModal();
+async function MostrarAlerta() {
+    const { value: formValues } = await Swal.fire({
+        title: "Iniciar sesion",
+        html: `
+      <input placeholder="Usuario" id="swal-input1" class="swal2-input">
+      <input type="password" id="swal-input2" class="swal2-input" placeholder="Contraseña">
+    `,
+        focusConfirm: false,
+        preConfirm: () => {
+            return [
+                document.getElementById("swal-input1").value,
+                document.getElementById("swal-input2").value
+            ];
+        }
     });
-})
+    if (formValues) {
+        const verificar = USUARIOS.some(persona => persona.usuario === formValues[0] && persona.contrasenia === formValues[1]); //VALIDAR DATOS DE INICIO
+        if (verificar) {
+            console.log("Bienvenido/a " + formValues[0]);
+            localStorage.setItem("logCorrecto", true);
+            window.location.href = "/index.html";
+        } else {
+            Swal.fire({ //CARTEL DE ERROR
+                icon: 'error',
+                title: 'Oops...',
+                text: 'Usuario y/o contraseña erróneos, por favor, reintente.'
+            });
+        }
+    }
+}
+
+const INICIAR = document.getElementById("iniciar-sesion");
+INICIAR.addEventListener("click", MostrarAlerta);
